\name{denormalize}
\alias{denormalize}
\title{Denormalize nested list structures into a flat data.frame}
\arguments{
  \item{o}{A list representation of a JSON object}

  \item{keep}{Only output these columns}

  \item{drop}{Output everything but these columns}
}
\description{
  Denormalize nested list structures into a flat data.frame
}

