\docType{package}
\name{odessa-package}
\alias{odessa}
\alias{odessa-package}
\title{A library that uses graph inference to effortlessly join datasets together}
\description{
  Odessa is a platform and library that eliminates the
  headaches associated with data cleaning and munging. With
  Odessa you can focus on the task of analyzing and
  modeling data and not wrestling with data.
}
\details{
  \tabular{ll}{ Package: \tab odessa\cr Type: \tab
  Package\cr Version: \tab 1.0.0\cr Date: \tab
  2013-10-14\cr License: \tab LGPL-3\cr LazyLoad: \tab
  yes\cr }
}
\author{
  Brian Lee Yung Rowe <r@zatonovo.com>
}
\seealso{
  \code{lambda.tools}
}
\keyword{package}

