\name{PackageDescriptor}
\alias{PackageDescriptor}
\title{A type that represents a data package}
\arguments{
  \item{id}{The Odessa ID for the dataset}
}
\value{
  A typed data.frame
}
\description{
  Every dataset in Odessa is a package containing the
  actual dataset and a binding file containing metadata.
}
\section{Usage}{
  PackageDescriptor(id)
}

\section{Details}{
  This is the CKAN package information
}
\author{
  Brian Lee Yung Rowe
}

