\name{set_binding}
\alias{set_binding}
\title{Update the binding file for a package}
\arguments{
  \item{id}{The Odessa ID for the dataset}

  \item{binding}{A data.frame that represents the binding
  data}
}
\description{
  When developing a new data package, this is useful for
  verifying that a binding file works correctly by updating
  a binding file locally.
}
\section{Usage}{
  set_binding(id, binding)
}

\section{Details}{
  Only use for dataset development
}
\author{
  Brian Lee Yung Rowe
}

