\name{which.ancestors}
\alias{which.ancestors}
\title{Check if any bindings can be derived via the odessa graph}
\arguments{
  \item{a}{An Odessa dataset}
}
\value{
  A character vector with all available ancestors
}
\description{
  Check if any bindings can be derived via the odessa graph
}
\author{
  Brian Lee Yung Rowe
}

