\name{which.bindings}
\alias{which.bindings}
\title{Find the available bindings for a data package}
\arguments{
  \item{a}{An Odessa dataset}
}
\value{
  A character vector with all available bindings
}
\description{
  Find the available bindings for a data package
}
\author{
  Brian Lee Yung Rowe
}

